Move get_manager to base module
