+++ pygeoapi/openapi.py
@@ -46,7 +46,7 @@ import yaml
 from pygeoapi import l10n
 from pygeoapi.models.openapi import OAPIFormat
 from pygeoapi.plugin import load_plugin
-from pygeoapi.process.manager import get_manager
+from pygeoapi.process.manager.base import get_manager
 from pygeoapi.provider.base import ProviderTypeError, SchemaType
 from pygeoapi.util import (filter_dict_by_key_value, get_provider_by_type,
                            filter_providers_by_type, to_json, yaml_load,
